requirements.txt
pytest
uuid
selenium
python-dotenv
dataclasses-json
